+++ doc/tutorial/read_mapper/read_mapper_step2.cpp
@@ -48,7 +48,7 @@ void map_reads(std::filesystem::path const & query_path,
     seqan3::sequence_file_input query_in{query_path};
 
     seqan3::configuration const search_config = seqan3::search_cfg::max_error{seqan3::search_cfg::total{errors}} |
-                                                seqan3::search_cfg::mode{seqan3::search_cfg::all_best};
+                                                seqan3::search_cfg::hit_all_best;
 
     for (auto & [query, id, qual] : query_in | seqan3::views::take(20))
     {