+++ doc/tutorial/read_mapper/read_mapper_step4.cpp
@@ -59,7 +59,7 @@ void map_reads(std::filesystem::path const & query_path,
 //! [alignment_file_output]
 
     seqan3::configuration const search_config = seqan3::search_cfg::max_error{seqan3::search_cfg::total{errors}} |
-                                                seqan3::search_cfg::mode{seqan3::search_cfg::all_best};
+                                                seqan3::search_cfg::hit_all_best;
 
     seqan3::configuration const align_config = seqan3::align_cfg::edit |
                                                seqan3::align_cfg::aligned_ends{seqan3::free_ends_first} |