+++ include/seqan3/search/detail/search_traits.hpp
@@ -15,7 +15,7 @@
 #include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/search/configuration/max_error.hpp>
 #include <seqan3/search/configuration/max_error_rate.hpp>
-#include <seqan3/search/configuration/mode.hpp>
+#include <seqan3/search/configuration/hit.hpp>
 #include <seqan3/search/configuration/output.hpp>
 
 namespace seqan3::detail
@@ -42,22 +42,18 @@ struct search_traits
     static constexpr bool has_error_configuration = search_with_max_error | search_with_max_error_rate;
 
     //!\brief A flag indicating whether search should find all hits.
-    static constexpr bool search_all_hits =
-        search_configuration_t::template exists<search_cfg::mode<detail::search_mode_all>>();
+    static constexpr bool search_all_hits = search_configuration_t::template exists<hit_all_tag>();
     //!\brief A flag indicating whether search should find best hits.
-    static constexpr bool search_best_hits =
-        search_configuration_t::template exists<search_cfg::mode<detail::search_mode_best>>();
+    static constexpr bool search_single_best_hit = search_configuration_t::template exists<hit_single_best_tag>();
     //!\brief A flag indicating whether search should find all best hits.
-    static constexpr bool search_all_best_hits =
-        search_configuration_t::template exists<search_cfg::mode<detail::search_mode_all_best>>();
+    static constexpr bool search_all_best_hits = search_configuration_t::template exists<hit_all_best_tag>();
     //!\brief A flag indicating whether search should find strata hits.
-    static constexpr bool search_strata_hits =
-        search_configuration_t::template exists<search_cfg::mode<search_cfg::strata>>();
-    //!\brief A flag indicating whether mode configuration was set in the search configuration.
-    static constexpr bool has_mode_configuration = search_all_hits |
-                                                   search_best_hits |
-                                                   search_all_best_hits |
-                                                   search_strata_hits;
+    static constexpr bool search_strata_hits = search_configuration_t::template exists<search_cfg::hit_strata>();
+    //!\brief A flag indicating whether hit configuration was set in the search configuration.
+    static constexpr bool has_hit_configuration = search_all_hits |
+                                                  search_single_best_hit |
+                                                  search_all_best_hits |
+                                                  search_strata_hits;
 
     //!\brief A flag indicating whether search should return the index cursor.
     static constexpr bool search_return_index_cursor =